Rename _schemas.ts to schemas.ts across vehicles routes
The underscore prefix was a misguided convention — SvelteKit only
treats files starting with `+` as routes, so any other name is a
regular module. Removed the underscore everywhere:

- vehicles/schemas.ts
- vehicles/[id]/reminders/schemas.ts
- vehicles/[id]/tasks/schemas.ts

Updated import paths in the four page.server.ts files that reference
them.
